+++ e2e/helpers/services/members-import/index.ts
@@ -1,1 +1,1 @@
-export * from './MembersImportService';
+export * from './members-import-service';
